Adding scrapeServices to remove responsibility of app.js
